refactor: make display layer truly read-only by removing lock coupling
**Problem:** display/data.js was checking locks and returning lock status,
creating backwards coupling (display telling indexing about indexing state).

**Solution:**
- Removed lock checking from loadMediaSheet (display layer)
- Moved all lock logic to indexing/locks.js (indexing layer)
- Coordinator (indexing) now checks locks separately
- Display layer is now purely read-only (just loads index files)

**Changes:**
- display/data.js: 295→228 lines (removed lock checks, saveMediaSheet)
- indexing/locks.js: Added checkIndexLock, isFreshIndexLock functions
- coordinator.js: Imports lock functions from locks.js, not display

**Clean separation achieved:**
```
display/data.js → reads index (no lock awareness)
indexing/locks.js → manages locks (write + read)
coordinator.js → checks locks, loads display data separately
```

All 580 tests passing ✓
